[Phase 6 Step 0] Relocate scenario types from jeod_runner to jeod_sim
Pure relocation, no behaviour change. Moves the per-vehicle config
types and the declarative simulation builder from `jeod_runner` to
`jeod_sim` so the standalone runner and the future Bevy adapter can
consume the same scenario library.

Types moved (each into its own module in `jeod_sim/src/`):

| Type                    | Source                          | New home               |
|-------------------------|---------------------------------|------------------------|
| `GravitySourceEntry`    | `jeod_runner/src/lib.rs`        | `sources.rs`           |
| `VehicleConfig`         | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `SrpModel`              | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `ShadowBody`            | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `GeodeticConfig`        | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `EarthLightingConfig`   | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `DerivedStateConfig`    | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `FrameSwitchConfig`     | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `SwitchSense`           | `jeod_runner/src/lib.rs`        | `vehicle_config.rs`    |
| `SimulationBuilder`     | `jeod_runner/src/builder.rs`    | `simulation_builder.rs` |
| `MassTreeAttachment`    | `jeod_runner/src/builder.rs`    | `simulation_builder.rs` |

Types that **stay** in `jeod_runner` (runtime concepts):

- `Simulation` — runtime state + step loop
- `SimBody` — per-body runtime state with integrator scratch
- `VehicleOutput` — runtime read-only snapshot
- `VehicleBuilder` (the runtime fluent builder) — Step 1 will
  consolidate it into the Phase-5 typestate `VehicleBuilder` in
  `jeod_sim`.

Terminal-method support: the orphan rule prevents `jeod_runner` from
adding inherent methods to `jeod_sim::SimulationBuilder`. Resolved via
two paths so call sites read identically to the pre-relocation API:

- `jeod_runner::Simulation::from_builder(b: SimulationBuilder) -> Result<Self, _>`
  — explicit constructor.
- `jeod_runner::SimulationBuilderExt::build` / `build_unchecked`
  — extension trait imported via `jeod_runner::prelude::*`.

`SimulationBuilder`'s fields are `pub` so consumer crates can
materialize without separate accessors. Pre-release status (#58)
makes this exposure acceptable; Phase 10 may re-encapsulate.

Compatibility: `jeod_runner::lib.rs` re-exports every relocated type,
so external `use jeod_runner::{VehicleConfig, GravitySourceEntry, …}`
imports continue to compile. Phase 10 will remove the re-exports.

`VehicleConfig::builder(trans)` had to become
`VehicleBuilder::new(trans)` — the same orphan-rule constraint
prevents `impl VehicleConfig` from living in `jeod_runner` now that
`VehicleConfig` moved. No external callers needed updating (the
constructor was only invoked from a doc-comment example).

## Verification

- `cargo fmt --check` clean
- `cargo clippy --workspace --tests -- -D warnings` clean
- `cargo nextest run --workspace -E 'not test(tier3_)'` — 722 pass
- `cargo nextest run -p jeod_runner --test tier3_sim_dyncomp_run2
  --test tier3_sim_dyncomp_run3` — 4/4 pass
- `tier3_baseline_diff` — 76 matched, 0 violations (bit-identical
  numerics, no kernel changes)
